+++ lib/jekyll_dynamic_assets.rb
@@ -4,6 +4,7 @@ require_relative "jekyll_dynamic_assets/version"
 require_relative "jekyll_dynamic_assets/processor"
 require_relative "jekyll_dynamic_assets/assets_tag"
 require_relative "jekyll_dynamic_assets/defaults/formats"
+require_relative "jekyll_dynamic_assets/defaults/sources"
 
 module JekyllDynamicAssets
   class Error < StandardError; end