+++ lib/jekyll_dynamic_assets/defaults/sources.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module JekyllDynamicAssets
+  DEFAULT_SOURCES = {
+    "base" => "/assets",
+    "css" => "/styles",
+    "js" => "/scripts",
+    "mjs" => "/scripts",
+    "ts" => "/scripts",
+    "json" => "/data",
+    "ico" => "/icons",
+    "woff" => "/fonts",
+    "woff2" => "/fonts",
+    "ttf" => "/fonts",
+    "otf" => "/fonts"
+  }.freeze
+end