+++ CHANGELOG.md
@@ -28,3 +28,20 @@
 ## [1.2.0] - 2025-06-06
 
 - Added option to select formats
+
+## [1.3.0] - 2025-06-07
+
+- Added Sub directory options
+- Added Default Sub directories for
+  - css   >> styles
+  - js    >> scripts
+  - mjs   >> scripts
+  - ts    >> scripts
+  - json  >> data
+  - ico   >> icons
+  - woff  >> fonts
+  - woff2 >> fonts
+  - ttf   >> fonts
+  - otf   >> fonts
+- Updated Selection logic, added inline definition
+- Added Absolute URL